quasar2/requirements.txt
absl-py==2.1.0
astropy==6.1.4
astropy-iers-data==0.2024.10.7.0.32.46
astunparse==1.6.3
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
dm-tree==0.1.8
docstring_parser==0.16
easyocr==1.7.1
etils==1.9.4
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gast==0.6.0
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
h5py==3.11.0
idna==3.8
imageio==2.35.1
immutabledict==4.2.0
importlib_resources==6.4.4
Jinja2==3.1.4
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.0
mpmath==1.3.0
namex==0.0.8
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
photoutils==0.0.1
photutils==1.13.0
pillow==10.4.0
promise==2.3
protobuf==4.25.4
psutil==6.0.0
pyarrow==17.0.0
pyclipper==1.3.0.post5
pyerfa==2.0.1.4
Pygments==2.18.0
pyparsing==3.1.4
python-bidi==0.6.0
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
rich==13.8.0
scikit-image==0.24.0
scipy==1.14.1
setuptools==74.1.2
shapely==2.0.6
simple_parsing==0.1.5
six==1.16.0
sympy==1.13.2
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.6
tensorflow-metadata==1.15.0
termcolor==2.4.0
tifffile==2024.8.30
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.1
